Demonstrate a projectile in an environment
